first moving sprites
DLExporter
- exporting custom background struct to hold clean background and tilemap for each room

Game
- renamed DISABLE_BACKGROUND_DRAWING to CUSTOM_BACKGROUND_DRAWING
- added custom struct to hold clean background and tileset for Resources
- added drops drawing
- changed dl_u32 to be long instead of int
- clean up dl_rand.*

+++ platforms/mastersystem/source/game/base_types.h
@@ -7,8 +7,8 @@ typedef unsigned char dl_u8;
 typedef signed short dl_s16;
 typedef unsigned short dl_u16;
 
-typedef signed int dl_s32;
-typedef unsigned int dl_u32;
+typedef signed long dl_s32;
+typedef unsigned long dl_u32;
 
 #define BOOL dl_u8
 #define TRUE 1
